Memory optimisations and text message update

+++ Parsers/VAP.cs
@@ -104,7 +104,7 @@ namespace IPTComShark.Parsers
                 new BitField()
                 {
                     Name = "First Niblet",
-                    BitFieldType = BitFieldType.UInt16,
+                    BitFieldType = BitFieldType.UInt8,
                     Length = 4
                 },
                 new BitField()
